% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apolloEngine.R
\name{ApolloEngine}
\alias{ApolloEngine}
\title{Business logic of the Apollo application}
\description{
An R6 object with methods for use in the Shiny application \code{apollo}.
}
\section{Super class}{
\code{\link[apollor:NA]{apollor::NA}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApolloEngine$new()}}
\item \href{#method-read_signals}{\code{ApolloEngine$read_signals()}}
\item \href{#method-read_actions}{\code{ApolloEngine$read_actions()}}
\item \href{#method-read_indicators}{\code{ApolloEngine$read_indicators()}}
\item \href{#method-read_persoon}{\code{ApolloEngine$read_persoon()}}
\item \href{#method-read_bedrijf}{\code{ApolloEngine$read_bedrijf()}}
\item \href{#method-read_adres}{\code{ApolloEngine$read_adres()}}
\item \href{#method-read_favourites}{\code{ApolloEngine$read_favourites()}}
\item \href{#method-read_log}{\code{ApolloEngine$read_log()}}
\item \href{#method-read_data_actualiteit}{\code{ApolloEngine$read_data_actualiteit()}}
\item \href{#method-list_actions}{\code{ApolloEngine$list_actions()}}
\item \href{#method-list_favourites}{\code{ApolloEngine$list_favourites()}}
\item \href{#method-get_log_ping}{\code{ApolloEngine$get_log_ping()}}
\item \href{#method-log_user_event}{\code{ApolloEngine$log_user_event()}}
\item \href{#method-add_favourite}{\code{ApolloEngine$add_favourite()}}
\item \href{#method-remove_favourite}{\code{ApolloEngine$remove_favourite()}}
\item \href{#method-create_action}{\code{ApolloEngine$create_action()}}
\item \href{#method-update_action}{\code{ApolloEngine$update_action()}}
\item \href{#method-archive_action}{\code{ApolloEngine$archive_action()}}
\item \href{#method-get_persoon_details}{\code{ApolloEngine$get_persoon_details()}}
\item \href{#method-get_adres_details}{\code{ApolloEngine$get_adres_details()}}
\item \href{#method-get_bedrijf_details}{\code{ApolloEngine$get_bedrijf_details()}}
\item \href{#method-clone}{\code{ApolloEngine$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="append_data">}\href{../../apollor/html/NA.html#method-append_data}{\code{apollor::NA$append_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="close">}\href{../../apollor/html/NA.html#method-close}{\code{apollor::NA$close()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="connect_to_database">}\href{../../apollor/html/NA.html#method-connect_to_database}{\code{apollor::NA$connect_to_database()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="delete_rows_where">}\href{../../apollor/html/NA.html#method-delete_rows_where}{\code{apollor::NA$delete_rows_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="execute_query">}\href{../../apollor/html/NA.html#method-execute_query}{\code{apollor::NA$execute_query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="has_value">}\href{../../apollor/html/NA.html#method-has_value}{\code{apollor::NA$has_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="have_column">}\href{../../apollor/html/NA.html#method-have_column}{\code{apollor::NA$have_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="list_tables">}\href{../../apollor/html/NA.html#method-list_tables}{\code{apollor::NA$list_tables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="make_column">}\href{../../apollor/html/NA.html#method-make_column}{\code{apollor::NA$make_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="query">}\href{../../apollor/html/NA.html#method-query}{\code{apollor::NA$query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="read_spatial_table">}\href{../../apollor/html/NA.html#method-read_spatial_table}{\code{apollor::NA$read_spatial_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="read_table">}\href{../../apollor/html/NA.html#method-read_table}{\code{apollor::NA$read_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where">}\href{../../apollor/html/NA.html#method-replace_value_where}{\code{apollor::NA$replace_value_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where2">}\href{../../apollor/html/NA.html#method-replace_value_where2}{\code{apollor::NA$replace_value_where2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_columns">}\href{../../apollor/html/NA.html#method-table_columns}{\code{apollor::NA$table_columns()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_info">}\href{../../apollor/html/NA.html#method-table_info}{\code{apollor::NA$table_info()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$new(gemeente, schema, pool, config_file = "conf/config.yml")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_signals"></a>}}
\if{latex}{\out{\hypertarget{method-read_signals}{}}}
\subsection{Method \code{read_signals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_signals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_actions"></a>}}
\if{latex}{\out{\hypertarget{method-read_actions}{}}}
\subsection{Method \code{read_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_actions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_indicators"></a>}}
\if{latex}{\out{\hypertarget{method-read_indicators}{}}}
\subsection{Method \code{read_indicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_indicators()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_persoon"></a>}}
\if{latex}{\out{\hypertarget{method-read_persoon}{}}}
\subsection{Method \code{read_persoon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_persoon()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_bedrijf"></a>}}
\if{latex}{\out{\hypertarget{method-read_bedrijf}{}}}
\subsection{Method \code{read_bedrijf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_bedrijf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_adres"></a>}}
\if{latex}{\out{\hypertarget{method-read_adres}{}}}
\subsection{Method \code{read_adres()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_adres()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_favourites"></a>}}
\if{latex}{\out{\hypertarget{method-read_favourites}{}}}
\subsection{Method \code{read_favourites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_favourites()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_log"></a>}}
\if{latex}{\out{\hypertarget{method-read_log}{}}}
\subsection{Method \code{read_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_data_actualiteit"></a>}}
\if{latex}{\out{\hypertarget{method-read_data_actualiteit}{}}}
\subsection{Method \code{read_data_actualiteit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_data_actualiteit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_actions"></a>}}
\if{latex}{\out{\hypertarget{method-list_actions}{}}}
\subsection{Method \code{list_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_actions(update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_favourites"></a>}}
\if{latex}{\out{\hypertarget{method-list_favourites}{}}}
\subsection{Method \code{list_favourites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_favourites(user = NULL, update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_ping"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_ping}{}}}
\subsection{Method \code{get_log_ping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_log_ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_user_event"></a>}}
\if{latex}{\out{\hypertarget{method-log_user_event}{}}}
\subsection{Method \code{log_user_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$log_user_event(username, action)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_favourite"></a>}}
\if{latex}{\out{\hypertarget{method-add_favourite}{}}}
\subsection{Method \code{add_favourite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_favourite(username, oid, object_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_favourite"></a>}}
\if{latex}{\out{\hypertarget{method-remove_favourite}{}}}
\subsection{Method \code{remove_favourite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$remove_favourite(username, favo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_action"></a>}}
\if{latex}{\out{\hypertarget{method-create_action}{}}}
\subsection{Method \code{create_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_action(
  registratie_id,
  username,
  actie_naam,
  datum_actie,
  omschrijving,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_action"></a>}}
\if{latex}{\out{\hypertarget{method-update_action}{}}}
\subsection{Method \code{update_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$update_action(
  action_id,
  actie_naam,
  registratie_id,
  username,
  datum_actie,
  omschrijving,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive_action"></a>}}
\if{latex}{\out{\hypertarget{method-archive_action}{}}}
\subsection{Method \code{archive_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$archive_action(action_id, username, actie_naam = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_persoon_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_persoon_details}{}}}
\subsection{Method \code{get_persoon_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_persoon_details(bsn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_adres_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_adres_details}{}}}
\subsection{Method \code{get_adres_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_adres_details(addreseerbaarobject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bedrijf_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_bedrijf_details}{}}}
\subsection{Method \code{get_bedrijf_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_bedrijf_details(kvknummer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
