% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralPretrainedModelTools.R
\docType{package}
\name{ApolloR-package}
\alias{ApolloR}
\alias{ApolloR-package}
\title{ApolloR: APOLLO R Package}
\description{
The R companion of Assessment of Pre-trained Observational Large Longitudinal models in OHDSI (APOLLO).
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
