% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tenant.R
\name{appid_tenant}
\alias{appid_tenant}
\title{Get app id (posit connect uuid) for a tenant}
\usage{
appid_tenant(tenant)
}
\arguments{
\item{tenant}{Juno tenant - not vectorized!}
}
\description{
Reads id from tenant_list (new format).
}
