% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tenant.R
\name{appname_tenant}
\alias{appname_tenant}
\title{Get app name for a tenant}
\usage{
appname_tenant(tenant)
}
\arguments{
\item{tenant}{Juno tenant - not vectorized!}
}
\description{
Reads appname from tenant_list (new format).
}
