% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetDescriptives.R
\name{computeParquetDescriptives}
\alias{computeParquetDescriptives}
\title{Compute descriptive statistics for Parquet files}
\usage{
computeParquetDescriptives(folder)
}
\arguments{
\item{folder}{The folder on the local file system where the Parquet files were
written.}
}
\value{
Does not return anything. Is called for the side-effect of generating the two
CSV files.
}
\description{
Computes descriptive statistics for data extracted using the \code{\link[=extractCdmToParquet]{extractCdmToParquet()}}
function. It will produce two CSV tables in the \code{folder}:
\itemize{
\item \strong{TableDescriptives.csv} will contain the row count and, where applicable, the
person count per table.
\item \strong{ConceptDescriptives.csv} will contain, for each concept, the number of occurrences,
and the number of persons having that concept.
}
}
