% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLP.R
\name{createApolloFinetuner}
\alias{createApolloFinetuner}
\title{Apollo Finetuner}
\usage{
createApolloFinetuner(
  numEpochs = 1,
  numFreezeEpochs = 1,
  learningRate = 3e-04,
  weightDecay = 1e-05,
  batchSize = 2,
  predictionHead = "lstm",
  pretrainedModelFolder = "path/to/model",
  parquetRootFolder = "path/to/data/",
  personSequenceFolder = NULL,
  maxCores = 1,
  device = "cuda"
)
}
\arguments{
\item{numEpochs}{Number of epochs to train the model.}

\item{numFreezeEpochs}{Number of epochs to freeze the pretrained model.}

\item{learningRate}{Learning rate for the optimizer.}

\item{weightDecay}{Weight decay for the optimizer.}

\item{batchSize}{Batch size for training.}

\item{predictionHead}{The type of prediction head to use. Options are "lstm" and "linear".}

\item{pretrainedModelFolder}{The folder containing the pretrained model.}

\item{parquetRootFolder}{The folder where the CDM data was written by the \code{extractCdmToParquet()} function.}

\item{personSequenceFolder}{The folder where the person sequence data was written by the
\code{processCdmData()} function. If not specified \code{processCdmData()} will be called}

\item{maxCores}{The maximum number of CPU cores to use during fine-tuning.}

\item{device}{The device to use for fine-tuning. Options are "cuda:x" where x is a number or "cpu".}
}
\description{
Apollo Finetuner
}
