% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLP.R
\name{createApolloSimpleModel}
\alias{createApolloSimpleModel}
\title{Apollo Simple Model}
\usage{
createApolloSimpleModel(
  numEpochs = 1,
  learningRate = 3e-04,
  weightDecay = 1e-05,
  batchSize = 512,
  personSequenceFolder = NULL,
  maxCores = 1,
  device = "cuda",
  maxSequenceLength = 512L
)
}
\arguments{
\item{numEpochs}{Number of epochs to train the model.}

\item{learningRate}{Learning rate for the optimizer.}

\item{weightDecay}{Weight decay for the optimizer.}

\item{batchSize}{Batch size for training.}

\item{personSequenceFolder}{The folder where the person sequence data was written by the
\code{processCdmData()} function.}

\item{maxCores}{The maximum number of CPU cores to use during fine-tuning.}

\item{device}{The device to use for fine-tuning. Options are "cuda:x" where x is a number or "cpu".}
}
\description{
Apollo Simple Model
}
