% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateBuilder.R
\name{createCdmCovariateSettings}
\alias{createCdmCovariateSettings}
\title{Create CDM covariate settings}
\usage{
createCdmCovariateSettings(
  folder,
  windowStart = -365,
  windowEnd = 0,
  partitions = 10,
  analysisId = 999
)
}
\arguments{
\item{folder}{The folder on the local file system where the Parquet files
will be written.}

\item{windowStart}{The start of the window relative to the cohort start date
where CDM data will be extracted.}

\item{windowEnd}{The end of the window relative to the cohort start date
where CDM data will be extracted.}

\item{partitions}{The number of partitions to divide the data in.}

\item{analysisId}{The covariate analysis ID.}
}
\value{
An object of type \code{covariateSettings}, to be used with the \code{FeatureExtraction}
package.
}
\description{
Create covariate settings for extracting verbatim data from a subset of fields
of a subset of tables in the OMOP Common Data Model
}
