% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pretrain.R
\name{createModelSettings}
\alias{createModelSettings}
\title{Create model settings}
\usage{
createModelSettings(
  maxSequenceLength = 512,
  conceptEmbedding = TRUE,
  visitOrderEmbedding = TRUE,
  segmentEmbedding = TRUE,
  ageEmbedding = TRUE,
  dateEmbedding = TRUE,
  visitConceptEmbedding = TRUE,
  hiddenSize = 768,
  numAttentionHeads = 12,
  numHiddenLayers = 12,
  intermediateSize = 3072,
  hiddenAct = "gelu",
  embeddingCombinationMethod = "sum",
  hiddenDropoutProb = 0.1,
  attentionProbsDropoutProb = 0.1
)
}
\arguments{
\item{maxSequenceLength}{The maximum length of the input sequence.}

\item{conceptEmbedding}{Use concept embedding?}

\item{visitOrderEmbedding}{Use visit order embedding?}

\item{segmentEmbedding}{Use segement embedding?}

\item{ageEmbedding}{Use age embedding?}

\item{dateEmbedding}{Use date embedding?}

\item{visitConceptEmbedding}{Use visit concept embedding?}

\item{hiddenSize}{Size of hidden layers.}

\item{numAttentionHeads}{Number of attention heads.}

\item{numHiddenLayers}{Number of hidden layers.}

\item{intermediateSize}{Size of intermediate layers.}

\item{hiddenAct}{Type of activation function used in the intermediate layer. Can be "gelu" or "relu".}

\item{embeddingCombinationMethod}{Type of embedding combination method. Can be "sum" or "concat".}

\item{hiddenDropoutProb}{Dropout probability for the hidden layers.}

\item{attentionProbsDropoutProb}{Dropout probability for the attention layer.}
}
\value{
A model settings object.
}
\description{
Create model settings
}
