% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineTuningAndPrediction.R
\name{createTrainingSettings}
\alias{createTrainingSettings}
\title{Create model training settings}
\usage{
createTrainingSettings(
  trainFraction = 1,
  numEpochs = 100,
  numFreezeEpochs = 1,
  learningRate = 0.001,
  weightDecay = 0,
  batchSize = 32
)
}
\arguments{
\item{trainFraction}{Fraction of the data to use for training. Set to 1 to skip evaluation.}

\item{numEpochs}{Number of epochs to train for.}

\item{numFreezeEpochs}{Number of epochs to freeze the pre-trained model for.}

\item{learningRate}{Learning rate for the Adam optimizer.}

\item{weightDecay}{Weight decay for the Adam optimizer.}
}
\value{
Training settings.
}
\description{
Create model training settings
}
