% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_functions.R
\name{get_tenant}
\alias{get_tenant}
\alias{open_config_site}
\alias{deploy}
\alias{deploy_test}
\title{Open the config_site for the current tenant}
\usage{
get_tenant()

open_config_site()

deploy(tenant = NULL, test = FALSE)

deploy_test(...)
}
\arguments{
\item{tenant}{If not provided, reads 'gemeente' from this_version}
}
\description{
Open config_site for current gemeente
}
