% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_all.R
\name{deploy_all}
\alias{deploy_all}
\title{Vectorized and safe version of deploy_now}
\usage{
deploy_all(
  tenants = NULL,
  where = "development",
  skip_tenants = NULL,
  confirm = TRUE,
  ...
)
}
\arguments{
\item{tenants}{vector of tenants to deploy (if missing, will deploy everything!)}

\item{where}{Refers to config section (development/production/)}

\item{skip_tenants}{Remove these tenants from the deploy list (for whatever reason)}

\item{confirm}{If no tenants argument given, prompt to continue unless confirm = FALSE}

\item{...}{Further arguments to \code{\link[=deploy_now]{deploy_now()}}, for example \code{test=TRUE}}
}
\description{
If no 'tenants' argument given, will deploy all tenants in the tenant list. If where='production',
only deploys those tenants with production: yes in the tenant list. Also cecks the db config file and prompts for missing passwords,
to avoid deploying a tenant without a database configuration.
}
