% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_now.R
\name{deploy_now}
\alias{deploy_now}
\title{Deploy Apollo directly}
\usage{
deploy_now(
  tenant,
  appname = NULL,
  where = "development",
  appid = NULL,
  appid_from_tenant_list = TRUE,
  test = FALSE,
  db_config_file = "conf/config.yml",
  log_deployment = TRUE,
  launch_browser = FALSE,
  deploy_location = NULL,
  delete_after_deploy = TRUE,
  ...
)
}
\arguments{
\item{tenant}{Tenant name (e.g. 'DEMO')}

\item{appname}{Name of app on posit connect - if NULL (the default), read from tenant_list.yml}

\item{where}{The server to deploy to (devapp or app)}

\item{appid}{Optional: the app ID (looks like a uuid in the url at posit connect). Needed if you
are a collaborator on an app and want to update it (the original uploader does not need the ID)}

\item{appid_from_tenant_list}{If TRUE (the default) and appid is NULL (also default), reads appid from the tenant
list if the deployment is to the production environment}

\item{test}{If TRUE, deploys to appname with suffix "_test", ignores appid}

\item{db_config_file}{Standard location of the DB config file}

\item{log_deployment}{If TRUE, attempts to log the deployment in DB}

\item{launch_browser}{If TRUE, opens a browser with the app after deployment}

\item{deploy_location}{App will be deployed from a temp directory, or set a directory here if
you want to inspect the pre-deploy files.}

\item{delete_after_deploy}{If TRUE, deletes all files after deploying.}

\item{...}{Further arguments passed to \code{\link[rsconnect:deployApp]{rsconnect::deployApp()}}}
}
\description{
Does not make a 'deploy project' but deploys Apollo for a tenant immediately,
from a temporary directory
}
