% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLP.R
\name{finetune}
\alias{finetune}
\title{Apollo Finetuner}
\usage{
finetune(trainData, modelSettings, analysisId, analysisPath, ...)
}
\arguments{
\item{trainData}{The covariate data to use for fine-tuning the model.}

\item{modelSettings}{The model settings to use for fine-tuning the model.}

\item{analysisId}{The analysis ID to use for fine-tuning the model.}

\item{analysisPath}{The path to save the analysis to.}
}
\description{
Apollo Finetuner
}
