% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineTuningAndPrediction.R
\name{fineTuneModel}
\alias{fineTuneModel}
\title{Fine-tune a pretrained model}
\usage{
fineTuneModel(
  pretrainedModelFolder,
  fineTunedModelFolder,
  covariateData,
  labels,
  trainingSettings = createTrainingSettings(),
  maxCores = 5
)
}
\arguments{
\item{pretrainedModelFolder}{The folder containing the pretrained model.}

\item{fineTunedModelFolder}{The folder where the fine-tuned model will be written.}

\item{covariateData}{The \code{CovariateData} object containing the CDM  data covariates.}

\item{labels}{The labels to use for training. A data frame containing at least two
columns: \code{rowId} and \code{outcomeCount}.}

\item{trainingSettings}{Parameters used when training the model.}

\item{maxCores}{The maximum number of CPU cores to use during fine-tuning. If a GPU
is found (CUDA or MPS), it will automatically be used irrespective
of the  setting of \code{maxCores}.}
}
\value{
Does not return anything. Is called for the side effect of creating the fine-tuned model and
saving it in the \code{fineTuneModel} folder.
}
\description{
Fine-tune a pretrained model
}
