% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLP.R
\name{finetunePredict}
\alias{finetunePredict}
\title{Apollo Finetuner prediction function}
\usage{
finetunePredict(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{The model to use for prediction.}

\item{data}{The covariate data to use for prediction.}

\item{cohort}{The cohort to use for prediction.}
}
\description{
Apollo Finetuner prediction function
}
