% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralPretrainedModelTools.R
\docType{package}
\name{GeneralPretrainedModelTools-package}
\alias{GeneralPretrainedModelTools}
\alias{GeneralPretrainedModelTools-package}
\title{GeneralPretrainedModelTools: Tools to Support General Pre-Trained Models from CDM Data}
\description{
Tools to help create large general pre-trained models from data in the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
