% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server.R
\name{get_server}
\alias{get_server}
\title{Posit connect server config}
\usage{
get_server(
  where = "development",
  servers = c(development = "devapp.shintolabs.net", production = "app.shintolabs.net",
    eindhoven_premium = "eindhoven.shintolabs.net")
)
}
\description{
Posit connect server config
}
