% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_functions.R
\name{get_tenant_choices}
\alias{get_tenant_choices}
\title{Get available tenants for Apollo}
\usage{
get_tenant_choices(path = getwd())
}
\arguments{
\item{path}{Path where to read the tenant_list (normally this working directory)}
}
\description{
Get available tenants for Apollo
}
