% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tenant.R
\name{has_production_tenant}
\alias{has_production_tenant}
\title{Decide whether to deploy to production or not}
\usage{
has_production_tenant(tenant)
}
\arguments{
\item{tenant}{Juno tenant - not vectorized!}
}
\description{
Reads id from tenant_list (new format).
}
