% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tenant.R
\name{label_tenant}
\alias{label_tenant}
\title{Label a tenant based on the code}
\usage{
label_tenant(tenant)
}
\arguments{
\item{tenant}{Tenant (e.g. 'DEMO'), can be vectorized}
}
\description{
Label a tenant based on the code
}
