% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apolloEngine.R
\name{LinkItEngine}
\alias{LinkItEngine}
\title{Business logic of the Apollo application}
\description{
An R6 object with methods for use in the Shiny application \code{apollo}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LinkItEngine$new()}}
\item \href{#method-list_tables}{\code{LinkItEngine$list_tables()}}
\item \href{#method-close}{\code{LinkItEngine$close()}}
\item \href{#method-read_signals}{\code{LinkItEngine$read_signals()}}
\item \href{#method-clone}{\code{LinkItEngine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$new(gemeente, schema, pool, config_file = "conf/config.yml")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$list_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_signals"></a>}}
\if{latex}{\out{\hypertarget{method-read_signals}{}}}
\subsection{Method \code{read_signals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$read_signals(user = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
