% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineTuningAndPrediction.R
\name{predictFineTuned}
\alias{predictFineTuned}
\title{Use a fine-tuned model to predict}
\usage{
predictFineTuned(fineTunedModelFolder, covariateData, population, maxCores = 5)
}
\arguments{
\item{fineTunedModelFolder}{The folder where the fine-tuned model was written by \code{fineTuneModel()}.}

\item{covariateData}{The \code{CovariateData} object containing the CDM data covariates.}

\item{population}{The population predict for. A data frame with at least one columns:
\code{rowId} and \code{outcomeCount}.}

\item{maxCores}{The maximum number of CPU cores to use during ine-tuning. If a GPU
is found (CUDA or MPS), it ill automatically be used irrespective of
the setting of \code{maxCores}.}
}
\value{
Returns a data frame with the set of row IDs in \code{population}, and a columnm \code{prediction}
containing the probability of belonging to the class.
}
\description{
Use a fine-tuned model to predict
}
