% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pretrain.R
\name{pretrainModel}
\alias{pretrainModel}
\title{Pretrain a model}
\usage{
pretrainModel(
  parquetFolder,
  pretrainedModelFolder,
  mappingSettings = createMappingSettings(),
  modelSettings = createModelSettings(),
  maxCores = 5
)
}
\arguments{
\item{parquetFolder}{The folder where the CDM data was written by the
\code{extractCdmToParquet()} function.}

\item{pretrainedModelFolder}{The folder containing the pretrained model.}

\item{mappingSettings}{Settings mapping CDM data to sequence data.}

\item{modelSettings}{Model settings as created using \code{createModelSettings()}.}

\item{maxCores}{The maximum number of CPU cores to use during pretraining. If a GPU
is found (CUDA or MPS), it will automatically be used irrespective
of the setting of \code{maxCores}.}
}
\value{
Does not return anything. Is called for the side effect of creating the pretrained model and
saving it in the \code{pretrainedModelFolder} folder.
}
\description{
Pretrain a model
}
\details{
Note: primarily intended for debugging purposes. I would expect you'd pretrain the model in Python.
}
