% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbi_key}
\alias{sbi_key}
\title{KvK SBI codes}
\format{
A data frame with 1434 rows and 8 variables:
\describe{
\item{sbi_code_txt}{}
\item{sbi_code_int}{}
\item{sbi_omschrijving}{}
\item{sbi_code_main}{}
\item{sbi_code_sub_1}{}
\item{sbi_code_sub_2}{}
\item{sbi_code_type}{}
\item{afdeling}{}
}
}
\usage{
sbi_key
}
\description{
KvK SBI codes
}
\keyword{datasets}
