% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbi_mapping}
\alias{sbi_mapping}
\title{KvK SBI mapping}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\usage{
sbi_mapping
}
\description{
A data frame with 21 rows and 2 variables:
\describe{
\item{afdeling}{}
\item{afdeling_omschrijving}{}
}
}
\keyword{datasets}
