% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_functions.R
\name{set_tenant}
\alias{set_tenant}
\title{Set current tenant in this_version.yml}
\usage{
set_tenant(tenant, path = getwd())
}
\arguments{
\item{tenant}{Tenant (e.g. 'DEMO')}

\item{path}{Path where to update the this_version.yml file}
}
\description{
Overwrites current this_version.yml!
}
