% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apolloEngine.R
\name{ApolloEngine}
\alias{ApolloEngine}
\title{Business logic of the Apollo application}
\description{
An R6 object with methods for use in the Shiny application \code{apollo}.
}
\section{Super class}{
\code{\link[apollor:NA]{apollor::NA}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApolloEngine$new()}}
\item \href{#method-assert_geo}{\code{ApolloEngine$assert_geo()}}
\item \href{#method-geo_name_from_code}{\code{ApolloEngine$geo_name_from_code()}}
\item \href{#method-add_geo_columns}{\code{ApolloEngine$add_geo_columns()}}
\item \href{#method-to_json}{\code{ApolloEngine$to_json()}}
\item \href{#method-from_json}{\code{ApolloEngine$from_json()}}
\item \href{#method-read_signals}{\code{ApolloEngine$read_signals()}}
\item \href{#method-read_actions}{\code{ApolloEngine$read_actions()}}
\item \href{#method-read_indicators}{\code{ApolloEngine$read_indicators()}}
\item \href{#method-read_person}{\code{ApolloEngine$read_person()}}
\item \href{#method-read_business}{\code{ApolloEngine$read_business()}}
\item \href{#method-read_address}{\code{ApolloEngine$read_address()}}
\item \href{#method-read_favorites}{\code{ApolloEngine$read_favorites()}}
\item \href{#method-read_log}{\code{ApolloEngine$read_log()}}
\item \href{#method-read_metadata}{\code{ApolloEngine$read_metadata()}}
\item \href{#method-add_indicator}{\code{ApolloEngine$add_indicator()}}
\item \href{#method-remove_indicator}{\code{ApolloEngine$remove_indicator()}}
\item \href{#method-get_indicators_theme}{\code{ApolloEngine$get_indicators_theme()}}
\item \href{#method-make_boolean_indicator}{\code{ApolloEngine$make_boolean_indicator()}}
\item \href{#method-make_indicator_table}{\code{ApolloEngine$make_indicator_table()}}
\item \href{#method-list_actions}{\code{ApolloEngine$list_actions()}}
\item \href{#method-list_favorites}{\code{ApolloEngine$list_favorites()}}
\item \href{#method-get_log_ping}{\code{ApolloEngine$get_log_ping()}}
\item \href{#method-log_user_event}{\code{ApolloEngine$log_user_event()}}
\item \href{#method-add_favorite}{\code{ApolloEngine$add_favorite()}}
\item \href{#method-remove_favorite}{\code{ApolloEngine$remove_favorite()}}
\item \href{#method-create_action}{\code{ApolloEngine$create_action()}}
\item \href{#method-update_action}{\code{ApolloEngine$update_action()}}
\item \href{#method-archive_action}{\code{ApolloEngine$archive_action()}}
\item \href{#method-get_person_from_id}{\code{ApolloEngine$get_person_from_id()}}
\item \href{#method-get_tags_for_person}{\code{ApolloEngine$get_tags_for_person()}}
\item \href{#method-get_address_from_id}{\code{ApolloEngine$get_address_from_id()}}
\item \href{#method-get_address_details}{\code{ApolloEngine$get_address_details()}}
\item \href{#method-get_business_details}{\code{ApolloEngine$get_business_details()}}
\item \href{#method-clone}{\code{ApolloEngine$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="append_data">}\href{../../apollor/html/NA.html#method-append_data}{\code{apollor::NA$append_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="close">}\href{../../apollor/html/NA.html#method-close}{\code{apollor::NA$close()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="connect_to_database">}\href{../../apollor/html/NA.html#method-connect_to_database}{\code{apollor::NA$connect_to_database()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="delete_rows_where">}\href{../../apollor/html/NA.html#method-delete_rows_where}{\code{apollor::NA$delete_rows_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="execute_query">}\href{../../apollor/html/NA.html#method-execute_query}{\code{apollor::NA$execute_query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="has_value">}\href{../../apollor/html/NA.html#method-has_value}{\code{apollor::NA$has_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="have_column">}\href{../../apollor/html/NA.html#method-have_column}{\code{apollor::NA$have_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="list_tables">}\href{../../apollor/html/NA.html#method-list_tables}{\code{apollor::NA$list_tables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="make_column">}\href{../../apollor/html/NA.html#method-make_column}{\code{apollor::NA$make_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="query">}\href{../../apollor/html/NA.html#method-query}{\code{apollor::NA$query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="read_spatial_table">}\href{../../apollor/html/NA.html#method-read_spatial_table}{\code{apollor::NA$read_spatial_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="read_table">}\href{../../apollor/html/NA.html#method-read_table}{\code{apollor::NA$read_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where">}\href{../../apollor/html/NA.html#method-replace_value_where}{\code{apollor::NA$replace_value_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where2">}\href{../../apollor/html/NA.html#method-replace_value_where2}{\code{apollor::NA$replace_value_where2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_columns">}\href{../../apollor/html/NA.html#method-table_columns}{\code{apollor::NA$table_columns()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_info">}\href{../../apollor/html/NA.html#method-table_info}{\code{apollor::NA$table_info()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$new(
  gemeente,
  schema,
  pool,
  config_file = getOption("config_file", "conf/config.yml"),
  geo_file = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_geo"></a>}}
\if{latex}{\out{\hypertarget{method-assert_geo}{}}}
\subsection{Method \code{assert_geo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$assert_geo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-geo_name_from_code"></a>}}
\if{latex}{\out{\hypertarget{method-geo_name_from_code}{}}}
\subsection{Method \code{geo_name_from_code()}}{
Translate buurt or wijk codes into names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$geo_name_from_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{E.g. "WK022801" or "BU02280105"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_geo_columns"></a>}}
\if{latex}{\out{\hypertarget{method-add_geo_columns}{}}}
\subsection{Method \code{add_geo_columns()}}{
Add wijk, gemeente columns to a dataframe with buurt_code_cbs column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_geo_columns(data)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Existing geo columns will be overwritten
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$to_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_signals"></a>}}
\if{latex}{\out{\hypertarget{method-read_signals}{}}}
\subsection{Method \code{read_signals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_signals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_actions"></a>}}
\if{latex}{\out{\hypertarget{method-read_actions}{}}}
\subsection{Method \code{read_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_actions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_indicators"></a>}}
\if{latex}{\out{\hypertarget{method-read_indicators}{}}}
\subsection{Method \code{read_indicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_indicators()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_person"></a>}}
\if{latex}{\out{\hypertarget{method-read_person}{}}}
\subsection{Method \code{read_person()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_person()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_business"></a>}}
\if{latex}{\out{\hypertarget{method-read_business}{}}}
\subsection{Method \code{read_business()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_business()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_address"></a>}}
\if{latex}{\out{\hypertarget{method-read_address}{}}}
\subsection{Method \code{read_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-read_favorites}{}}}
\subsection{Method \code{read_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_favorites()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_log"></a>}}
\if{latex}{\out{\hypertarget{method-read_log}{}}}
\subsection{Method \code{read_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-read_metadata}{}}}
\subsection{Method \code{read_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-add_indicator}{}}}
\subsection{Method \code{add_indicator()}}{
Adds an indicator's metadata to the 'indicator' table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_indicator(
  indicator_name,
  type = c("address", "person"),
  label,
  description,
  user_id,
  theme,
  columns,
  weight,
  threshold
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Indicator type - refers to the table that has this column_name}

\item{\code{label}}{Short label for the indicator}

\item{\code{description}}{Long-form text explaining the indicator}

\item{\code{user_id}}{Who are you?}

\item{\code{theme}}{Which theme(s) is the indicator used in? e.g. c("mensenhandel","drugs")}

\item{\code{columns}}{}

\item{\code{column_name}}{Name of the column (indicator)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-remove_indicator}{}}}
\subsection{Method \code{remove_indicator()}}{
Remove an indicator's metadata from the 'indicator' table0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$remove_indicator(indicator_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_indicators_theme"></a>}}
\if{latex}{\out{\hypertarget{method-get_indicators_theme}{}}}
\subsection{Method \code{get_indicators_theme()}}{
Get rows of indicators table for a theme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_indicators_theme(theme)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use this function to find definitions for indicators in a theme.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_boolean_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-make_boolean_indicator}{}}}
\subsection{Method \code{make_boolean_indicator()}}{
Convert raw indicator data to TRUE/FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$make_boolean_indicator(data, indicator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe subset from \code{indicator} table, read with \verb{$get_indicators_theme}}

\item{\code{indicator}}{Name of the indicator to convert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_indicator_table"></a>}}
\if{latex}{\out{\hypertarget{method-make_indicator_table}{}}}
\subsection{Method \code{make_indicator_table()}}{
Make a table with indicator (boolean) values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$make_indicator_table(
  theme,
  type = c("address", "person", "business"),
  id_columns = c("address_id", "buurt_code_cbs")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{One of the themes in the \code{indicator} table}

\item{\code{type}}{One of the \link{object_type}'s in the \code{indicator} table}

\item{\code{id_columns}}{Columns from the base table (e.g. person) that are copied to the indicator table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_actions"></a>}}
\if{latex}{\out{\hypertarget{method-list_actions}{}}}
\subsection{Method \code{list_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_actions(update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-list_favorites}{}}}
\subsection{Method \code{list_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_favorites(user = NULL, update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_ping"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_ping}{}}}
\subsection{Method \code{get_log_ping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_log_ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_user_event"></a>}}
\if{latex}{\out{\hypertarget{method-log_user_event}{}}}
\subsection{Method \code{log_user_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$log_user_event(user_id, description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-add_favorite}{}}}
\subsection{Method \code{add_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_favorite(user_id, object_id, object_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-remove_favorite}{}}}
\subsection{Method \code{remove_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$remove_favorite(user_id, favorite_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_action"></a>}}
\if{latex}{\out{\hypertarget{method-create_action}{}}}
\subsection{Method \code{create_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_action(
  registration_id,
  user_id,
  action_name,
  action_date,
  description,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_action"></a>}}
\if{latex}{\out{\hypertarget{method-update_action}{}}}
\subsection{Method \code{update_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$update_action(
  action_id,
  action_name,
  registration_id,
  user_id,
  action_date,
  description,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive_action"></a>}}
\if{latex}{\out{\hypertarget{method-archive_action}{}}}
\subsection{Method \code{archive_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$archive_action(action_id, user_id, action_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_person_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_person_from_id}{}}}
\subsection{Method \code{get_person_from_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_person_from_id(person_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tags_for_person"></a>}}
\if{latex}{\out{\hypertarget{method-get_tags_for_person}{}}}
\subsection{Method \code{get_tags_for_person()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_tags_for_person(person_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_address_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_address_from_id}{}}}
\subsection{Method \code{get_address_from_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_address_from_id(address_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_address_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_address_details}{}}}
\subsection{Method \code{get_address_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_address_details(addreseerbaarobject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_business_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_business_details}{}}}
\subsection{Method \code{get_business_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_business_details(kvknummer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
