% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apolloEngine.R
\name{ApolloEngine}
\alias{ApolloEngine}
\title{Business logic of the Apollo application}
\description{
An R6 object with methods for use in the Shiny application \code{apollo}.
}
\section{Super class}{
\code{\link[apollor:NA]{apollor::NA}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApolloEngine$new()}}
\item \href{#method-assert_geo}{\code{ApolloEngine$assert_geo()}}
\item \href{#method-geo_name_from_code}{\code{ApolloEngine$geo_name_from_code()}}
\item \href{#method-add_geo_columns}{\code{ApolloEngine$add_geo_columns()}}
\item \href{#method-get_cbs_buurt_data}{\code{ApolloEngine$get_cbs_buurt_data()}}
\item \href{#method-get_cbs_buurt_metadata}{\code{ApolloEngine$get_cbs_buurt_metadata()}}
\item \href{#method-street_from_bagid}{\code{ApolloEngine$street_from_bagid()}}
\item \href{#method-get_bag_from_bagid}{\code{ApolloEngine$get_bag_from_bagid()}}
\item \href{#method-join_bag_geometry}{\code{ApolloEngine$join_bag_geometry()}}
\item \href{#method-to_json}{\code{ApolloEngine$to_json()}}
\item \href{#method-from_json}{\code{ApolloEngine$from_json()}}
\item \href{#method-read_signals}{\code{ApolloEngine$read_signals()}}
\item \href{#method-read_actions}{\code{ApolloEngine$read_actions()}}
\item \href{#method-read_indicator}{\code{ApolloEngine$read_indicator()}}
\item \href{#method-read_person}{\code{ApolloEngine$read_person()}}
\item \href{#method-read_metadata}{\code{ApolloEngine$read_metadata()}}
\item \href{#method-read_business}{\code{ApolloEngine$read_business()}}
\item \href{#method-read_address}{\code{ApolloEngine$read_address()}}
\item \href{#method-read_favorites}{\code{ApolloEngine$read_favorites()}}
\item \href{#method-read_log}{\code{ApolloEngine$read_log()}}
\item \href{#method-read_mpp}{\code{ApolloEngine$read_mpp()}}
\item \href{#method-get_indicator}{\code{ApolloEngine$get_indicator()}}
\item \href{#method-get_metadata}{\code{ApolloEngine$get_metadata()}}
\item \href{#method-set_metadata}{\code{ApolloEngine$set_metadata()}}
\item \href{#method-label_indicator}{\code{ApolloEngine$label_indicator()}}
\item \href{#method-add_indicator}{\code{ApolloEngine$add_indicator()}}
\item \href{#method-remove_indicator}{\code{ApolloEngine$remove_indicator()}}
\item \href{#method-get_indicators_theme}{\code{ApolloEngine$get_indicators_theme()}}
\item \href{#method-get_indicators_riskmodel}{\code{ApolloEngine$get_indicators_riskmodel()}}
\item \href{#method-make_boolean_indicator}{\code{ApolloEngine$make_boolean_indicator()}}
\item \href{#method-make_indicator_table}{\code{ApolloEngine$make_indicator_table()}}
\item \href{#method-combine_indicator_tables}{\code{ApolloEngine$combine_indicator_tables()}}
\item \href{#method-calculate_riskmodel}{\code{ApolloEngine$calculate_riskmodel()}}
\item \href{#method-set_indicator_weight}{\code{ApolloEngine$set_indicator_weight()}}
\item \href{#method-write_indicator_riskmodel}{\code{ApolloEngine$write_indicator_riskmodel()}}
\item \href{#method-get_active_actions}{\code{ApolloEngine$get_active_actions()}}
\item \href{#method-list_actions}{\code{ApolloEngine$list_actions()}}
\item \href{#method-list_favorites}{\code{ApolloEngine$list_favorites()}}
\item \href{#method-list_persons_by_id}{\code{ApolloEngine$list_persons_by_id()}}
\item \href{#method-list_addresses_by_id}{\code{ApolloEngine$list_addresses_by_id()}}
\item \href{#method-get_log_ping}{\code{ApolloEngine$get_log_ping()}}
\item \href{#method-log_user_event}{\code{ApolloEngine$log_user_event()}}
\item \href{#method-get_favorite}{\code{ApolloEngine$get_favorite()}}
\item \href{#method-add_favorite}{\code{ApolloEngine$add_favorite()}}
\item \href{#method-remove_favorite}{\code{ApolloEngine$remove_favorite()}}
\item \href{#method-create_MPP_for_registration}{\code{ApolloEngine$create_MPP_for_registration()}}
\item \href{#method-archive_MPP_for_registration_old}{\code{ApolloEngine$archive_MPP_for_registration_old()}}
\item \href{#method-archive_MPP_for_registration}{\code{ApolloEngine$archive_MPP_for_registration()}}
\item \href{#method-updateMPP}{\code{ApolloEngine$updateMPP()}}
\item \href{#method-update_MPP_for_registration}{\code{ApolloEngine$update_MPP_for_registration()}}
\item \href{#method-create_action}{\code{ApolloEngine$create_action()}}
\item \href{#method-update_action}{\code{ApolloEngine$update_action()}}
\item \href{#method-archive_action}{\code{ApolloEngine$archive_action()}}
\item \href{#method-get_person_from_id}{\code{ApolloEngine$get_person_from_id()}}
\item \href{#method-get_tags_for_person}{\code{ApolloEngine$get_tags_for_person()}}
\item \href{#method-get_address_from_id}{\code{ApolloEngine$get_address_from_id()}}
\item \href{#method-get_residents}{\code{ApolloEngine$get_residents()}}
\item \href{#method-get_businesses_at_address}{\code{ApolloEngine$get_businesses_at_address()}}
\item \href{#method-get_relocations_for_person}{\code{ApolloEngine$get_relocations_for_person()}}
\item \href{#method-get_relocations_timeline}{\code{ApolloEngine$get_relocations_timeline()}}
\item \href{#method-create_network_nodes}{\code{ApolloEngine$create_network_nodes()}}
\item \href{#method-create_network_edges}{\code{ApolloEngine$create_network_edges()}}
\item \href{#method-clone}{\code{ApolloEngine$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="append_data">}\href{../../apollor/html/NA.html#method-append_data}{\code{apollor::NA$append_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="close">}\href{../../apollor/html/NA.html#method-close}{\code{apollor::NA$close()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="connect_to_database">}\href{../../apollor/html/NA.html#method-connect_to_database}{\code{apollor::NA$connect_to_database()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="delete_rows_where">}\href{../../apollor/html/NA.html#method-delete_rows_where}{\code{apollor::NA$delete_rows_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="execute_query">}\href{../../apollor/html/NA.html#method-execute_query}{\code{apollor::NA$execute_query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="has_value">}\href{../../apollor/html/NA.html#method-has_value}{\code{apollor::NA$has_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="have_column">}\href{../../apollor/html/NA.html#method-have_column}{\code{apollor::NA$have_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="list_tables">}\href{../../apollor/html/NA.html#method-list_tables}{\code{apollor::NA$list_tables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="make_column">}\href{../../apollor/html/NA.html#method-make_column}{\code{apollor::NA$make_column()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="query">}\href{../../apollor/html/NA.html#method-query}{\code{apollor::NA$query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="read_table">}\href{../../apollor/html/NA.html#method-read_table}{\code{apollor::NA$read_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where">}\href{../../apollor/html/NA.html#method-replace_value_where}{\code{apollor::NA$replace_value_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="replace_value_where2">}\href{../../apollor/html/NA.html#method-replace_value_where2}{\code{apollor::NA$replace_value_where2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_columns">}\href{../../apollor/html/NA.html#method-table_columns}{\code{apollor::NA$table_columns()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="table_info">}\href{../../apollor/html/NA.html#method-table_info}{\code{apollor::NA$table_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="apollor" data-topic="NA" data-id="to_sql_string">}\href{../../apollor/html/NA.html#method-to_sql_string}{\code{apollor::NA$to_sql_string()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$new(
  gemeente,
  schema,
  pool,
  config_file = getOption("config_file", "conf/config.yml"),
  geo_file = NULL,
  load_data = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_geo"></a>}}
\if{latex}{\out{\hypertarget{method-assert_geo}{}}}
\subsection{Method \code{assert_geo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$assert_geo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-geo_name_from_code"></a>}}
\if{latex}{\out{\hypertarget{method-geo_name_from_code}{}}}
\subsection{Method \code{geo_name_from_code()}}{
Translate buurt or wijk codes into names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$geo_name_from_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{E.g. "WK022801" or "BU02280105"
TODO this is not vectorized!}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_geo_columns"></a>}}
\if{latex}{\out{\hypertarget{method-add_geo_columns}{}}}
\subsection{Method \code{add_geo_columns()}}{
Add wijk, gemeente columns to a dataframe with buurt_code_cbs column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_geo_columns(data, spatial = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Existing geo columns will be overwritten
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cbs_buurt_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_cbs_buurt_data}{}}}
\subsection{Method \code{get_cbs_buurt_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_cbs_buurt_data(buurt_code_cbs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cbs_buurt_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_cbs_buurt_metadata}{}}}
\subsection{Method \code{get_cbs_buurt_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_cbs_buurt_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-street_from_bagid"></a>}}
\if{latex}{\out{\hypertarget{method-street_from_bagid}{}}}
\subsection{Method \code{street_from_bagid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$street_from_bagid(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bag_from_bagid"></a>}}
\if{latex}{\out{\hypertarget{method-get_bag_from_bagid}{}}}
\subsection{Method \code{get_bag_from_bagid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_bag_from_bagid(id, spatial = FALSE, geo_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-join_bag_geometry"></a>}}
\if{latex}{\out{\hypertarget{method-join_bag_geometry}{}}}
\subsection{Method \code{join_bag_geometry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$join_bag_geometry(data, id_column = "adresseerbaarobject_id")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$to_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_signals"></a>}}
\if{latex}{\out{\hypertarget{method-read_signals}{}}}
\subsection{Method \code{read_signals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_signals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_actions"></a>}}
\if{latex}{\out{\hypertarget{method-read_actions}{}}}
\subsection{Method \code{read_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_actions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-read_indicator}{}}}
\subsection{Method \code{read_indicator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_indicator()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_person"></a>}}
\if{latex}{\out{\hypertarget{method-read_person}{}}}
\subsection{Method \code{read_person()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_person()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-read_metadata}{}}}
\subsection{Method \code{read_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_business"></a>}}
\if{latex}{\out{\hypertarget{method-read_business}{}}}
\subsection{Method \code{read_business()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_business()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_address"></a>}}
\if{latex}{\out{\hypertarget{method-read_address}{}}}
\subsection{Method \code{read_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-read_favorites}{}}}
\subsection{Method \code{read_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_favorites(user_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_log"></a>}}
\if{latex}{\out{\hypertarget{method-read_log}{}}}
\subsection{Method \code{read_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_mpp"></a>}}
\if{latex}{\out{\hypertarget{method-read_mpp}{}}}
\subsection{Method \code{read_mpp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$read_mpp(registration_id = NULL, archived = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-get_indicator}{}}}
\subsection{Method \code{get_indicator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_indicator(indicator_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$set_metadata(
  name,
  label,
  timestamp_provided,
  owner,
  depends_on,
  step,
  colnames,
  description,
  timestamp_processed = Sys.time()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-label_indicator}{}}}
\subsection{Method \code{label_indicator()}}{
Get labels for indicators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$label_indicator(indicator_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indicator_name}}{Vector of indicator names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-add_indicator}{}}}
\subsection{Method \code{add_indicator()}}{
Adds an indicator's metadata to the 'indicator' table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_indicator(
  indicator_name,
  type = c("address", "person"),
  label,
  description,
  user_id,
  theme,
  columns
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Indicator type - refers to the table that has this column_name}

\item{\code{label}}{Short label for the indicator}

\item{\code{description}}{Long-form text explaining the indicator}

\item{\code{user_id}}{Who are you?}

\item{\code{theme}}{Which theme(s) is the indicator used in? e.g. c("mensenhandel","drugs")}

\item{\code{columns}}{}

\item{\code{column_name}}{Name of the column (indicator)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-remove_indicator}{}}}
\subsection{Method \code{remove_indicator()}}{
Remove an indicator's metadata from the 'indicator' table0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$remove_indicator(indicator_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_indicators_theme"></a>}}
\if{latex}{\out{\hypertarget{method-get_indicators_theme}{}}}
\subsection{Method \code{get_indicators_theme()}}{
Get rows of indicators table for a theme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_indicators_theme(theme)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use this function to find definitions for indicators in a theme.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_indicators_riskmodel"></a>}}
\if{latex}{\out{\hypertarget{method-get_indicators_riskmodel}{}}}
\subsection{Method \code{get_indicators_riskmodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_indicators_riskmodel(user_id, theme, gemeente)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_boolean_indicator"></a>}}
\if{latex}{\out{\hypertarget{method-make_boolean_indicator}{}}}
\subsection{Method \code{make_boolean_indicator()}}{
Convert raw indicator data to TRUE/FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$make_boolean_indicator(data, indicator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe subset from \code{indicator} table, read with \verb{$get_indicators_theme}}

\item{\code{indicator}}{Name of the indicator to convert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_indicator_table"></a>}}
\if{latex}{\out{\hypertarget{method-make_indicator_table}{}}}
\subsection{Method \code{make_indicator_table()}}{
Make a table with indicator (boolean) values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$make_indicator_table(
  theme,
  type = c("address", "person", "business"),
  id_columns = c("address_id", "buurt_code_cbs"),
  user_id,
  gemeente
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{One of the themes in the \code{indicator} table}

\item{\code{type}}{One of the \link{object_type}'s in the \code{indicator} table}

\item{\code{id_columns}}{Columns from the base table (e.g. person) that are copied to the indicator table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-combine_indicator_tables"></a>}}
\if{latex}{\out{\hypertarget{method-combine_indicator_tables}{}}}
\subsection{Method \code{combine_indicator_tables()}}{
Combine address level and person level indicators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$combine_indicator_tables(indi_address, indi_person)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indi_address}}{Indicators at address level}

\item{\code{indi_person}}{Indicators at person level}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make an address-level dataset with indicators. The person indicator table
is summarized so that if at least one person at the address is TRUE for some indicator,
then the address is TRUE. TODO business level indicators
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_riskmodel"></a>}}
\if{latex}{\out{\hypertarget{method-calculate_riskmodel}{}}}
\subsection{Method \code{calculate_riskmodel()}}{
Calculate riskmodel at address level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$calculate_riskmodel(data, theme, user_id, gemeente)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Combined indicator table at address level (made with \code{combine_indicator_tables})}

\item{\code{theme}}{Theme for the indicators; used to get weights from definition table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_indicator_weight"></a>}}
\if{latex}{\out{\hypertarget{method-set_indicator_weight}{}}}
\subsection{Method \code{set_indicator_weight()}}{
Set the weight (riskmodel) for an indicator in a theme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$set_indicator_weight(indicator_name, theme, weight)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_indicator_riskmodel"></a>}}
\if{latex}{\out{\hypertarget{method-write_indicator_riskmodel}{}}}
\subsection{Method \code{write_indicator_riskmodel()}}{
Write weight/threshold for an indicator in a theme for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$write_indicator_riskmodel(
  user_id,
  indicator_name,
  theme,
  weight,
  threshold
)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Written to table indicator_riskmodel. If the entry does not exist for the user,
add a new row for this user. If it does exist, use an UPDATE statement to replace the value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_active_actions"></a>}}
\if{latex}{\out{\hypertarget{method-get_active_actions}{}}}
\subsection{Method \code{get_active_actions()}}{
Get the active actions ordered by date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_active_actions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_actions"></a>}}
\if{latex}{\out{\hypertarget{method-list_actions}{}}}
\subsection{Method \code{list_actions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_actions(update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-list_favorites}{}}}
\subsection{Method \code{list_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_favorites(user_id = NULL, update = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_persons_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-list_persons_by_id}{}}}
\subsection{Method \code{list_persons_by_id()}}{
Get persons based on person or address_id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_persons_by_id(address_id = NULL, person_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{address_id}}{Vector of address IDs}

\item{\code{person_id}}{Vector of person Ids. Provide either address or person ID, not both!}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_addresses_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-list_addresses_by_id}{}}}
\subsection{Method \code{list_addresses_by_id()}}{
Get addresses based on address_id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$list_addresses_by_id(address_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{address_id}}{Vector of address IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_ping"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_ping}{}}}
\subsection{Method \code{get_log_ping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_log_ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_user_event"></a>}}
\if{latex}{\out{\hypertarget{method-log_user_event}{}}}
\subsection{Method \code{log_user_event()}}{
Log an user action
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$log_user_event(user_id, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User that performed action}

\item{\code{description}}{The action perfomed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-get_favorite}{}}}
\subsection{Method \code{get_favorite()}}{
check if object is currently in favorites of user_id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_favorite(user_id, object_id, object_type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User to check for}

\item{\code{object_id}}{Identifier of the object}

\item{\code{object_type}}{Type of object (address, person, business or registration)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-add_favorite}{}}}
\subsection{Method \code{add_favorite()}}{
Add favorite to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$add_favorite(user_id, object_id, object_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User to add to}

\item{\code{object_id}}{Identifier of the object}

\item{\code{object_type}}{Type of object (address, person, business or registration)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Favorites can be addresses, persons, businesses and registrations
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-remove_favorite}{}}}
\subsection{Method \code{remove_favorite()}}{
Remove from favorites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$remove_favorite(user_id, favorite_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User to remove from}

\item{\code{favorite_id}}{Identifier of the removable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_MPP_for_registration"></a>}}
\if{latex}{\out{\hypertarget{method-create_MPP_for_registration}{}}}
\subsection{Method \code{create_MPP_for_registration()}}{
Create MPP for registration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_MPP_for_registration(
  registration_id,
  user_id,
  data,
  createLog = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive_MPP_for_registration_old"></a>}}
\if{latex}{\out{\hypertarget{method-archive_MPP_for_registration_old}{}}}
\subsection{Method \code{archive_MPP_for_registration_old()}}{
Archive mpp for registration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$archive_MPP_for_registration_old(
  registration_id,
  user_id,
  createLog = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive_MPP_for_registration"></a>}}
\if{latex}{\out{\hypertarget{method-archive_MPP_for_registration}{}}}
\subsection{Method \code{archive_MPP_for_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$archive_MPP_for_registration(
  registration_id,
  user_id,
  mpp_names,
  createLog = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateMPP"></a>}}
\if{latex}{\out{\hypertarget{method-updateMPP}{}}}
\subsection{Method \code{updateMPP()}}{
Update mpp for registration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$updateMPP(registration_id, data, user_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_MPP_for_registration"></a>}}
\if{latex}{\out{\hypertarget{method-update_MPP_for_registration}{}}}
\subsection{Method \code{update_MPP_for_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$update_MPP_for_registration(
  registration_id,
  data,
  user_id,
  registration_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_action"></a>}}
\if{latex}{\out{\hypertarget{method-create_action}{}}}
\subsection{Method \code{create_action()}}{
Add action to actionlist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_action(
  registration_id,
  user_id,
  action_name,
  action_date,
  description,
  status
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{registration_id}}{Signal where the action relates to}

\item{\code{user_id}}{User that creates the action}

\item{\code{action_name}}{Short title}

\item{\code{action_date}}{Date (yyyy-mm-dd) action occurs}

\item{\code{description}}{Content of action}

\item{\code{status}}{Current action status}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_action"></a>}}
\if{latex}{\out{\hypertarget{method-update_action}{}}}
\subsection{Method \code{update_action()}}{
Update action in actionlist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$update_action(
  action_id,
  action_name,
  registration_id,
  user_id,
  action_date,
  description,
  status
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_name}}{Short title}

\item{\code{registration_id}}{Signal where the action relates to}

\item{\code{user_id}}{User that updates the action}

\item{\code{action_date}}{Date (yyyy-mm-dd) action occurs}

\item{\code{description}}{Content of action}

\item{\code{status}}{Current action status}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive_action"></a>}}
\if{latex}{\out{\hypertarget{method-archive_action}{}}}
\subsection{Method \code{archive_action()}}{
Archive action in actionlist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$archive_action(action_id, user_id, action_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{Action to archive}

\item{\code{user_id}}{User that archives the action}

\item{\code{action_name}}{\link{optional} Short title}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_person_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_person_from_id}{}}}
\subsection{Method \code{get_person_from_id()}}{
Get person details from the identifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_person_from_id(person_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person_id}}{Person's identifier FI: (pseudo)bsn}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tags_for_person"></a>}}
\if{latex}{\out{\hypertarget{method-get_tags_for_person}{}}}
\subsection{Method \code{get_tags_for_person()}}{
Hardcoded method to simulate tags
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_tags_for_person(person_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person_id}}{Person's identifier FI: (pseudo)bsn}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_address_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_address_from_id}{}}}
\subsection{Method \code{get_address_from_id()}}{
Get address details from the identifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_address_from_id(address_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{address_id}}{Address's identifier FI: (pseudo)adresseerbaarobject}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_residents"></a>}}
\if{latex}{\out{\hypertarget{method-get_residents}{}}}
\subsection{Method \code{get_residents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_residents(address_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_businesses_at_address"></a>}}
\if{latex}{\out{\hypertarget{method-get_businesses_at_address}{}}}
\subsection{Method \code{get_businesses_at_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_businesses_at_address(address_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_relocations_for_person"></a>}}
\if{latex}{\out{\hypertarget{method-get_relocations_for_person}{}}}
\subsection{Method \code{get_relocations_for_person()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_relocations_for_person(person_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_relocations_timeline"></a>}}
\if{latex}{\out{\hypertarget{method-get_relocations_timeline}{}}}
\subsection{Method \code{get_relocations_timeline()}}{
Create data suitable for timeline plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$get_relocations_timeline(person_id, add_overlijden = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person_id}}{Person's identifier FI: (pseudo)bsn}

\item{\code{add_overlijden}}{Boolean indicating if overlijden should be added as event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_network_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-create_network_nodes}{}}}
\subsection{Method \code{create_network_nodes()}}{
Create suitable node format for network(Viz)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_network_nodes(
  person_data = NULL,
  address_data = NULL,
  resident_data = NULL,
  business_data = NULL,
  registration_data = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_network_edges"></a>}}
\if{latex}{\out{\hypertarget{method-create_network_edges}{}}}
\subsection{Method \code{create_network_edges()}}{
Create suitable edge format for network(Viz)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$create_network_edges(
  person_data = NULL,
  address_data = NULL,
  resident_data = NULL,
  business_data = NULL,
  registration_data = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApolloEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
