% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pretrain.R
\name{createModelSettings}
\alias{createModelSettings}
\title{Create model settings}
\usage{
createModelSettings(
  maxSequenceLength = 512,
  conceptEmbedding = TRUE,
  visitOrderEmbedding = TRUE,
  segmentEmbedding = TRUE,
  ageEmbedding = TRUE,
  dateEmbedding = TRUE,
  visitConceptEmbedding = TRUE,
  hiddenSize = 768,
  numAttentionHeads = 12,
  numHiddenLayers = 12,
  intermediateSize = 3072,
  hiddenAct = "gelu",
  embeddingCombinationMethod = "sum",
  hiddenDropoutProb = 0.1,
  attentionProbsDropoutProb = 0.1
)
}
\arguments{
\item{maxSequenceLength}{}

\item{conceptEmbedding}{}

\item{visitOrderEmbedding}{}

\item{segmentEmbedding}{}

\item{ageEmbedding}{}

\item{dateEmbedding}{}

\item{visitConceptEmbedding}{}

\item{hiddenSize}{}

\item{numAttentionHeads}{}

\item{numHiddenLayers}{}

\item{intermediateSize}{}

\item{hiddenAct}{}

\item{embeddingCombinationMethod}{}

\item{hiddenDropoutProb}{}

\item{attentionProbsDropoutProb}{}
}
\value{
A model settings object.
}
\description{
Create model settings
}
