% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineTuningAndPrediction.R
\name{createTrainingSettings}
\alias{createTrainingSettings}
\title{Create model training settings}
\usage{
createTrainingSettings(
  trainFraction = 1,
  numEpochs = 100,
  numFreezeEpochs = 1,
  learningRate = 0.001,
  weightDecay = 0.01
)
}
\arguments{
\item{trainFraction}{}

\item{numEpochs}{}

\item{numFreezeEpochs}{}

\item{learningRate}{}

\item{weightDecay}{}
}
\description{
Create model training settings
}
