% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_functions.R
\name{set_tenant}
\alias{set_tenant}
\alias{get_tenant}
\alias{get_tenant_choices}
\alias{open_config_site}
\alias{deploy}
\title{Set tenant}
\usage{
set_tenant(tenant)

get_tenant()

get_tenant_choices(path = getwd())

open_config_site()

deploy(tenant = NULL)
}
\arguments{
\item{tenant}{If not provided, reads 'gemeente' from this_version}
}
\description{
Set current tenant ('gemeente') in this_version.yml. Overwrites current this_version.yml!

Get available gemeentes for the WBM

Open config_site for current gemeente
}
