% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExtraction.R
\name{extractCdmToParquet}
\alias{extractCdmToParquet}
\title{Extract data from the database}
\usage{
extractCdmToParquet(
  connectionDetails,
  cdmDatabaseSchema,
  workDatabaseSchema,
  partitionTablePrefix = "GPM_",
  folder,
  sampleSize = 1e+06,
  partitions = 200,
  maxCores = 3,
  forceRestart = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{workDatabaseSchema}{The name of the database schema where work tables can be created.}

\item{partitionTablePrefix}{The prefix to use when creating table names in the
\code{workDatabaseSChema} for storing the person ID and concept ID
partition tables.}

\item{folder}{The folder on the local file system where the Parquet files will
be written.}

\item{sampleSize}{The number of persons to be included in the sample.}

\item{partitions}{The number of partitions. Fewer partitions may lead to memory
issues.}

\item{maxCores}{The maximum number of parallel threads to use.}

\item{forceRestart}{If \code{FALSE}, when data is already found in the \code{folder} the process
will continue where it left off. If \code{TRUE}, any existing data files
will be deleted, and the process will start from scratch.}
}
\description{
Extract data from the server for a random sample of persons, and stores them
in the local file system as Parquet files. The data are split in to a number of
partitions (the number of partitions can be specified by the user).
}
