% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApolloR.R
\docType{package}
\name{ApolloR-package}
\alias{ApolloR}
\alias{ApolloR-package}
\title{ApolloR: Tools to Support General Pre-Trained Models from CDM Data}
\description{
Tools to help create large general pre-trained models from data in the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
